+++ config/post-type-list-table-options.php
@@ -1,12 +1,6 @@
 <?php
 return [
     [
-        'key' => 'post_type_key',
-        'label' => 'Post Type Key',
-        'description' => 'Typically singular version of word/concept; ie: book',
-        'type' => 'string'
-    ],
-    [
         'key' => 'post_type_singular',
         'label' => 'Post Type Singular',
         'description' => 'Singular version of word/content (no spaces), ie: Book',